+++ docs/en/changes/changes.md
@@ -46,6 +46,8 @@
 * Add protection for dynamic config change propagate chain.
 * Add Ruby component IDs.
 * Add component ID(160) for Caffeine.
+* Alarm: Support store and query the metrics snapshot when the alarm is triggered.
+* Alarm: Remove unused `Alarm Trend` query.
 * Fix missing remote endpoint IP address in span query of zipkin query module.
 
 #### UI